PetroliqController: - refactor to update doubles to decimals (one int).
UserSettings and UserSettingsForRegistration:
- refactor to update doubles to decimals.
- new properties to match db schema changes; RoundUnitCostTo, IdealDiscount, CurrentBatchId, NextFillId, AvgCapacityUnitPerDistanceUnit, MaxVolumeQualifyingForDiscount.

Update readme.

+++ Petroliq/Controllers/PetroliqController.cs
@@ -12,35 +12,35 @@ namespace Petroliq_API.Controllers
     public class PetroliqController : ControllerBase
     {
         private readonly IConfiguration _configuration;
-        private readonly double _kmToMiFactor; // Petroliq:KM_MI_FACTOR
-        private readonly double _miToKmFactor; // Petroliq:MI_KM_FACTOR
-        private readonly double _ltrToGalFactor; // Petroliq:LTR_GAL_FACTOR
-        private readonly double _defaultRoundTo; // Petroliq:DEFAULT_ROUND_TO
+        private readonly decimal _kmToMiFactor; // Petroliq:KM_MI_FACTOR
+        private readonly decimal _miToKmFactor; // Petroliq:MI_KM_FACTOR
+        private readonly decimal _ltrToGalFactor; // Petroliq:LTR_GAL_FACTOR
+        private readonly int _defaultRoundTo; // Petroliq:DEFAULT_ROUND_TO
         private readonly bool _petroliqSettingsParsed = false;
 
 #pragma warning disable CS1591
-        public double GetKmToMiFactor()
+        public decimal GetKmToMiFactor()
         {
             return _petroliqSettingsParsed ? _kmToMiFactor : -1;
         }
 #pragma warning restore CS1591
 
 #pragma warning disable CS1591
-        public double GetMiToKmFactor()
+        public decimal GetMiToKmFactor()
         {
             return _petroliqSettingsParsed ? _miToKmFactor : -1;
         }
 #pragma warning restore CS1591
 
 #pragma warning disable CS1591
-        public double GetLtrToGalFactor()
+        public decimal GetLtrToGalFactor()
         {
             return _petroliqSettingsParsed ? _ltrToGalFactor : -1;
         }
 #pragma warning restore CS1591
 
 #pragma warning disable CS1591
-        public double GetDefaultRoundTo()
+        public int GetDefaultRoundTo()
         {
             return _petroliqSettingsParsed ? _defaultRoundTo : -1;
         }
@@ -54,11 +54,11 @@ namespace Petroliq_API.Controllers
         {
             _configuration = configuration;
 
-            _petroliqSettingsParsed = 
-                double.TryParse(_configuration["Petroliq:KM_MI_FACTOR"], out _kmToMiFactor) &&            
-                double.TryParse(_configuration["Petroliq:MI_KM_FACTOR"], out _miToKmFactor) &&            
-                double.TryParse(_configuration["Petroliq:LTR_GAL_FACTOR"], out _ltrToGalFactor) &&            
-                double.TryParse(_configuration["Petroliq:DEFAULT_ROUND_TO"], out _defaultRoundTo);
+            _petroliqSettingsParsed =
+                decimal.TryParse(_configuration["Petroliq:KM_MI_FACTOR"], out _kmToMiFactor) &&
+                decimal.TryParse(_configuration["Petroliq:MI_KM_FACTOR"], out _miToKmFactor) &&
+                decimal.TryParse(_configuration["Petroliq:LTR_GAL_FACTOR"], out _ltrToGalFactor) &&            
+                int.TryParse(_configuration["Petroliq:DEFAULT_ROUND_TO"], out _defaultRoundTo);
         }
 
         /// <summary>
